+++ Application/Interface/IAuthenService.cs
@@ -1,0 +1,11 @@
+﻿using Infrastructure.Models;
+
+namespace Application.Interface
+{
+    public interface IAuthenService
+    {
+        User GetUserByUsername(string username);
+        bool VerifyPassword(string password, string passwordHash);
+        string GenerateJwtToken(User user);
+    }
+}
