+++ Application/Request/UpdateUserRequest.cs
@@ -1,0 +1,12 @@
+﻿namespace Application.Request
+{
+    public class UpdateUserRequest
+    {
+        public string UserName { get; set; }
+        public string Password { get; set; }
+        public string Name { get; set; }
+        public string Address { get; set; }
+        public string Phone { get; set; }
+        public string Email { get; set; }
+    }
+}
