+++ Application/Request/UserLogin.cs
@@ -1,0 +1,8 @@
+﻿namespace Application.Request
+{
+    public class UserLogin
+    {
+        public string Username { get; set; }
+        public string Password { get; set; }
+    }
+}
