+++ Application/Request/UsersFillterRequest.cs
@@ -1,0 +1,10 @@
+﻿namespace Application.Request
+{
+    public class UsersFillterRequest
+    {
+        public string Name { get; set; }
+        public string Address { get; set; }
+        public DateTime? CreatedDateFrom { get; set; }
+        public DateTime? CreatedDateTo { get; set; }
+    }
+}
