+++ Application/Response/CreateUserResponse.cs
@@ -1,0 +1,15 @@
+﻿namespace Application.Response
+{
+    public class CreateUserResponse
+    {
+        public string UserName { get; set; }
+
+        public string Name { get; set; }
+
+        public string Address { get; set; }
+
+        public string Email { get; set; }
+
+        public string Phone { get; set; }
+    }
+}
