+++ Application/Response/UpdateUserResponse.cs
@@ -1,0 +1,11 @@
+﻿namespace Application.Response
+{
+    public class UpdateUserResponse
+    {
+        public string Name { get; set; }
+
+        public string Address { get; set; }
+
+        public string Phone { get; set; }
+    }
+}
