+++ Application/Response/UsersFillterResponse.cs
@@ -1,0 +1,16 @@
+﻿namespace Application.Response
+{
+    public class UsersFillterResponse
+    {
+        public Guid Id { get; set; }
+        public string UserName { get; set; }
+        public string Name { get; set; }
+        public string Address { get; set; }
+        public string Email { get; set; }
+        public string Phone { get; set; }
+        public DateTime? CreatedDate { get; set; }
+        public string CreatedBy { get; set; }
+        public DateTime? ModifiedDate { get; set; }
+        public string ModifiedBy { get; set; }
+    }
+}
