--- Common/Helper/StringHashHelper.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Common.Helper
-{
-    public static class StringHashHelper
-    {
-        public static string PasswordHash(string password)
-        {
-            return password;
-        }
-    }
-}
