+++ Common/Helper/StringHelper.cs
@@ -1,0 +1,10 @@
+﻿namespace Common.Helper
+{
+    public static class StringHelper
+    {
+        public static string PasswordHash(string password)
+        {
+            return password;
+        }
+    }
+}
