+++ Common/IRepository/IRepository.cs
@@ -1,0 +1,12 @@
+﻿namespace Common.IRepository
+{
+    public interface IRepository<T> where T : class
+    {
+        Task<IEnumerable<T>> GetAllAsync();
+        Task<T?> GetByIdAsync(int id);
+        Task AddAsync(T entity);
+        void Update(T entity);
+        void Delete(T entity);
+    }
+
+}
