+++ Infrastructure/ApplicationDbContext/AppDbContext.cs
@@ -1,0 +1,12 @@
+﻿using Infrastructure.Models;
+using Microsoft.EntityFrameworkCore;
+
+namespace Infrastructure.ApplicationDbContext
+{
+    public class AppDbContext : DbContext
+    {
+        public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) { }
+
+        public DbSet<User> Users { get; set; }
+    }
+}
