+++ Infrastructure/Models/User.cs
@@ -1,0 +1,46 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Infrastructure.Models
+{
+    public class User
+    {
+        // Guid Id
+        public Guid Id { get; set; }
+        [StringLength(50)]
+        public string UserName { get; set; }
+
+        [StringLength(100)]
+        public string Password { get; set; }
+
+        [StringLength(20)]
+        public string GroupID { set; get; }
+
+        [StringLength(50)]
+        public string Name { get; set; }
+
+        [StringLength(50)]
+        public string Address { get; set; }
+
+        [StringLength(50)]
+        public string Email { get; set; }
+
+        [StringLength(50)]
+        public string Phone { get; set; }
+
+        public int? ProvinceID { set; get; }
+
+        public int? DistrictID { set; get; }
+
+        public DateTime? CreatedDate { get; set; }
+
+        [StringLength(50)]
+        public string CreatedBy { get; set; }
+
+        public DateTime? ModifiedDate { get; set; }
+
+        [StringLength(50)]
+        public string ModifiedBy { get; set; }
+
+        public bool Status { get; set; }
+    }
+}
