+++ UnitTest/UnitTest.csproj
@@ -7,8 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Middleware\" />
-    <Folder Include="IUnitOfWork\" />
+    <Folder Include="UserTest\" />
   </ItemGroup>
 
 </Project>
