--- UserManagement/Data/DbContext.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using UserManagement.Models.Entity;
-
-namespace UserManagement.Data
-{
-    public class AppDbContext : DbContext
-    {
-        public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) { }
-
-        public DbSet<User> Users { get; set; }
-    }
-}
