--- UserManagement/Models/ViewModels/Request/UsersFillterRequest.cs
@@ -1,10 +1,0 @@
-﻿namespace UserManagement.Models.ViewModels.Request
-{
-    public class UsersFillterRequest
-    {
-        public string Name { get; set; }
-        public string Address { get; set; }
-        public DateTime? CreatedDateFrom { get; set; }
-        public DateTime? CreatedDateTo { get; set; }
-    }
-}
