--- UserManagement/Models/ViewModels/Response/UpdateUserResponse.cs
@@ -1,11 +1,0 @@
-﻿namespace UserManagement.Models.ViewModels.Response
-{
-    public class UpdateUserResponse
-    {
-        public string Name { get; set; }
-
-        public string Address { get; set; }
-
-        public string Phone { get; set; }
-    }
-}
