--- UserManagement/Models/ViewModels/UserLogin.cs
@@ -1,8 +1,0 @@
-﻿namespace UserManagement.Models.ViewModels
-{
-    public class UserLogin
-    {
-        public string Username { get; set; }
-        public string Password { get; set; }
-    }
-}
