--- UserManagement/Services/Implement/AuthenService.cs
@@ -1,51 +1,0 @@
-﻿using Microsoft.IdentityModel.Tokens;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Security.Cryptography;
-using System.Text;
-using UserManagement.Data;
-using UserManagement.Models.Entity;
-using UserManagement.Services.Interface;
-
-namespace UserManagement.Services.Implement
-{
-    public class AuthenService : IAuthenService
-    {
-        private readonly AppDbContext _context;
-        private readonly IConfiguration _config;
-
-        public AuthenService(AppDbContext context, IConfiguration config)
-        {
-            _context = context;
-            _config = config;
-        }
-
-        public User GetUserByUsername(string username)
-        {
-            return _context.Users.FirstOrDefault(u => u.UserName == username);
-        }
-
-        public bool VerifyPassword(string password, string passwordHash)
-        {
-            using var sha256 = SHA256.Create();
-            var hash = Convert.ToBase64String(sha256.ComputeHash(Encoding.UTF8.GetBytes(password)));
-            return hash == passwordHash;
-        }
-
-        public string GenerateJwtToken(User user)
-        {
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var key = Encoding.ASCII.GetBytes(_config["Jwt:Key"]);
-            var tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Subject = new ClaimsIdentity(new[] {
-                new Claim(ClaimTypes.Name, user.UserName)
-            }),
-                Expires = DateTime.UtcNow.AddHours(1),
-                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
-            };
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            return tokenHandler.WriteToken(token);
-        }
-    }
-}
